components tsx 전환 및 typing 추가 Header, IssueCard, NoList
