--- src/components/Test.tsx
@@ -1,7 +1,0 @@
-import React from "react"
-
-type Props = {}
-
-export default function Test({}: Props) {
-	return <div>Test</div>
-}
